+++ problems/0006/solution.jl
@@ -1,7 +1,7 @@
 using BenchmarkTools
 BenchmarkTools.DEFAULT_PARAMETERS.samples = 100
 
-compute(n::Integer)::Integer = n * (n + 1) * (3 * n ^ 2 - n - 2) ÷ 12
+compute(n::Int64)::Int64 = n * (n + 1) * (3 * n ^ 2 - n - 2) ÷ 12
 
 compute(10)
