+++ problems/0017/compute.jl
@@ -1,4 +1,4 @@
-function compute(number)
+function compute(number::Int64)::Int64
     numbers_below_10 = [
         length("one"),
         length("two"),
